LPS-51413 avoid hard coding references and static classes
